+++ src/quantum/azext_quantum/vendored_sdks/azure_mgmt_quantum/models/quantum_workspace.py
@@ -20,13 +20,13 @@ class QuantumWorkspace(TrackedResource):
 
     All required parameters must be populated in order to send to Azure.
 
-    :ivar id: Fully qualified resource Id for the resource. Ex -
+    :ivar id: Fully qualified resource ID for the resource. Ex -
      /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
     :vartype id: str
     :ivar name: The name of the resource
     :vartype name: str
-    :ivar type: The type of the resource. Ex-
-     Microsoft.Compute/virtualMachines or Microsoft.Storage/storageAccounts.
+    :ivar type: The type of the resource. E.g.
+     "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
     :vartype type: str
     :param tags: Resource tags.
     :type tags: dict[str, str]
@@ -45,8 +45,12 @@ class QuantumWorkspace(TrackedResource):
     :param storage_account: ARM Resource Id of the storage account associated
      with this workspace.
     :type storage_account: str
+    :ivar endpoint_uri: The URI of the workspace endpoint.
+    :vartype endpoint_uri: str
     :param identity: Managed Identity information.
     :type identity: ~azure.quantum.models.QuantumWorkspaceIdentity
+    :ivar system_data: System metadata
+    :vartype system_data: ~azure.quantum.models.SystemData
     """
 
     _validation = {
@@ -56,6 +60,8 @@ class QuantumWorkspace(TrackedResource):
         'location': {'required': True},
         'usable': {'readonly': True},
         'provisioning_state': {'readonly': True},
+        'endpoint_uri': {'readonly': True},
+        'system_data': {'readonly': True},
     }
 
     _attribute_map = {
@@ -68,7 +74,9 @@ class QuantumWorkspace(TrackedResource):
         'usable': {'key': 'properties.usable', 'type': 'str'},
         'provisioning_state': {'key': 'properties.provisioningState', 'type': 'str'},
         'storage_account': {'key': 'properties.storageAccount', 'type': 'str'},
+        'endpoint_uri': {'key': 'properties.endpointUri', 'type': 'str'},
         'identity': {'key': 'identity', 'type': 'QuantumWorkspaceIdentity'},
+        'system_data': {'key': 'systemData', 'type': 'SystemData'},
     }
 
     def __init__(self, **kwargs):
@@ -77,4 +85,6 @@ class QuantumWorkspace(TrackedResource):
         self.usable = None
         self.provisioning_state = None
         self.storage_account = kwargs.get('storage_account', None)
+        self.endpoint_uri = None
         self.identity = kwargs.get('identity', None)
+        self.system_data = None